datatypes

+++ SQL/01_Loading/create_voronoi.sql
@@ -5,7 +5,7 @@ DROP TABLE IF EXISTS public.voronoi CASCADE;
 --Create table voronoi
 CREATE TABLE public.voronoi
 (
-  id numeric(10,0),
+  id bigint,
   geom geometry(Polygon,4326),
   CONSTRAINT voronoi_pkey PRIMARY KEY (id)
 )
